+++ apps/server/src/modules/learnroom/repo/index.ts
@@ -2,9 +2,9 @@ export { ColumnBoardBoardElement } from './mikro-orm/column-board-board-element.
 export { ColumnBoardNode, ColumnBoardNodeProps } from './mikro-orm/column-board-node.entity';
 export { ColumnBoardNodeRepo } from './mikro-orm/column-board-node.repo';
 export { CourseMikroOrmRepo } from './mikro-orm/course.repo';
+export { DashboardElementRepo } from './mikro-orm/dashboard-element.repo';
 export { DashboardModelMapper } from './mikro-orm/dashboard.model.mapper';
 export { DashboardRepo } from './mikro-orm/dashboard.repo';
-export { DashboardElementRepo } from './mikro-orm/dashboardElement.repo';
 export { LegacyBoardElement, LegacyBoardElementType } from './mikro-orm/legacy-board-element.entity';
 export { BoardProps, LegacyBoard } from './mikro-orm/legacy-board.entity';
 export { LegacyBoardRepo } from './mikro-orm/legacy-board.repo';
